Add landing page and added the general funcionality of the site
