+++ client/src/components/Hero/Hero.js
@@ -2,6 +2,10 @@
 
 import React from "react";
 import { Link } from "react-router-dom";
+
+import Register from "../Register/Register";
+import Background from "../Background/Background";
+
 import Typography from "@material-ui/core/Typography";
 import Button from "@material-ui/core/Button";
 
@@ -18,7 +22,12 @@ const Hero = ({ onRouteChange }) => {
       </Typography>
       <div id="buttonDiv">
         <Button id="register-now-btn" color="primary" variant="raised">
-          <Link to="/register">Register</Link>
+          <Link
+            to="/register"
+            style={{ textDecoration: "none", color: "#fff" }}
+          >
+            Register
+          </Link>
         </Button>
       </div>
     </div>
